norminette et memccpy

+++ ft_memchr.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *str, int c, int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *str, int c, size_t n)
 {
 	int		i;
